requirements.txt
fastapi>=0.110.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.29.0
telethon>=1.34.0,<1.35.0
jinja2>=3.1.2,<3.2.0
python-dotenv>=1.0.1,<2.0.0
